requirements.txt
streamlit
PyPDF2
google-genai 
chromadb
python-dotenv